+++ src/greynoise/api/filter.py
@@ -16,7 +16,7 @@ class Filter(object):
     def __init__(self, api):
         self.api = api
 
-    def filter(self, text, noise_only):
+    def filter(self, text, noise_only, riot_only):
         """Filter lines that contain IP addresses from a given text.
 
         :param text: Text input
@@ -25,6 +25,10 @@ class Filter(object):
             If set, return only lines that contain IP addresses classified as noise,
             otherwise, return lines that contain IP addresses not classified as noise.
         :type noise_only: bool
+        :param riot_only:
+            If set, return only lines that contain IP addresses in RIOT,
+            otherwise, return lines that contain IP addresses not in RIOT.
+        :type riot_only: bool
         :return: Iterator that yields lines in chunks
         :rtype: iterable
 
@@ -33,9 +37,9 @@ class Filter(object):
             text = text.splitlines(True)
         chunks = more_itertools.chunked(text, self.FILTER_TEXT_CHUNK_SIZE)
         for chunk in chunks:
-            yield self._filter_chunk(chunk, noise_only)
+            yield self._filter_chunk(chunk, noise_only, riot_only)
 
-    def _filter_chunk(self, text, noise_only):
+    def _filter_chunk(self, text, noise_only, riot_only):  # noqa: C901
         """Filter chunk of lines that contain IP addresses from a given text.
 
         :param text: Text input
@@ -44,6 +48,10 @@ class Filter(object):
             If set, return only lines that contain IP addresses classified as noise,
             otherwise, return lines that contain IP addresses not classified as noise.
         :type noise_only: bool
+        :param riot_only:
+            If set, return only lines that contain IP addresses in RIOT,
+            otherwise, return lines that contain IP addresses not in RIOT.
+        :type riot_only: bool
         :return: Filtered line
 
         """
@@ -51,11 +59,14 @@ class Filter(object):
         for input_line in text:
             text_ip_addresses.update(self.api.IPV4_REGEX.findall(input_line))
 
-        noise_ip_addresses = {
-            result["ip"]
-            for result in self.api.quick(text_ip_addresses)
-            if result["noise"]
-        }
+        noise_ip_addresses = []
+        riot_ip_addresses = []
+
+        for result in self.api.quick(text_ip_addresses):
+            if result["noise"]:
+                noise_ip_addresses.append(result["ip"])
+            if result["riot"]:
+                riot_ip_addresses.append(result["ip"])
 
         def all_ip_addresses_noisy(line):
             """Select lines that contain IP addresses and all of them are noisy.
@@ -72,6 +83,21 @@ class Filter(object):
                 for line_ip_address in line_ip_addresses
             )
 
+        def all_ip_addresses_riot(line):
+            """Select lines that contain IP addresses and all of them are in RIOT.
+
+            :param line: Line being processed.
+            :type line: str
+            :return: True if line contains IP addresses and all of them are noisy.
+            :rtype: bool
+
+            """
+            line_ip_addresses = self.api.IPV4_REGEX.findall(line)
+            return line_ip_addresses and all(
+                line_ip_address in riot_ip_addresses
+                for line_ip_address in line_ip_addresses
+            )
+
         def add_markup(match):
             """Add markup to surround IP address value with proper tag.
 
@@ -84,6 +110,8 @@ class Filter(object):
             ip_address = match.group(0)
             if ip_address in noise_ip_addresses:
                 tag = "noise"
+            elif ip_address in riot_ip_addresses:
+                tag = "riot"
             else:
                 tag = "not-noise"
 
@@ -91,6 +119,8 @@ class Filter(object):
 
         if noise_only:
             line_matches = all_ip_addresses_noisy
+        elif riot_only:
+            line_matches = all_ip_addresses_riot
         else:
 
             def line_matches(line):
@@ -102,7 +132,9 @@ class Filter(object):
                 :rtype: bool
 
                 """
-                return not all_ip_addresses_noisy(line)
+                return not all_ip_addresses_noisy(line) and not all_ip_addresses_riot(
+                    line
+                )
 
         filtered_lines = [
             self.api.IPV4_REGEX.subn(add_markup, input_line)[0]